+++ src/main/java/fastmessage/whatsapp/FastWhatsApp.java
@@ -1,11 +1,11 @@
-package fastmessaging.whatsapp;
+package FastMessage.whatsapp;
 
-import fastmessaging.ByteSlice;
-import fastmessaging.MessagePayload;
-import fastmessaging.MessageStatus;
-import fastmessaging.MessageType;
-import fastmessaging.MessagingChannel;
-import fastmessaging.UniversalMessage;
+import FastMessage.ByteSlice;
+import FastMessage.MessagePayload;
+import FastMessage.MessageStatus;
+import FastMessage.MessageType;
+import FastMessage.MessagingChannel;
+import FastMessage.UniversalMessage;
 
 import javax.crypto.Mac;
 import javax.crypto.spec.SecretKeySpec;